use fake fs for offline unit test
